Fix
Fix code seo create useSeo to get meta tags and clear main page
